+++ src/interfaces/ICheckout.ts
@@ -1,4 +1,4 @@
-import { ICheckoutSessions } from "./ICheckoutSession";
+import { ICheckoutSessions } from "./ICheckoutSessions";
 
 export interface ICheckout {
     sessions: ICheckoutSessions;
