+++ src/interfaces/index.ts
@@ -1,5 +1,5 @@
 import { ICheckout } from "./ICheckout";
-import { ICheckoutSessions } from "./ICheckoutSession";
+import { ICheckoutSessions } from "./ICheckoutSessions";
 import { IDiagonal } from "./IDiagonal";
 
 export { IDiagonal, ICheckout, ICheckoutSessions };
